+++ R/covmat.R
@@ -10,16 +10,19 @@
 #'
 #'   Family size is used for weighting covariance terms in case of multiple
 #'   half-sib families. It only matters if number of progeny differs.
-#' @param linkDam (p x p) matrix of maternal LD between pairs of p markers;
+#' @param linkMat (p x p) matrix of maternal LD between pairs of p markers;
 #'   matrix is block diagonal in case of multiple chromosomes
-#' @param haploSire (2N x p) matrix of sires haplotypes for all chromosomes
-#'   (2 lines per sire)
+#' @param haploMat (2N x p) matrix of sires haplotypes for all chromosomes
+#'   (2 lines per sire) codes as 0's and 1's reflecting reference and alternate
+#'   alleles
 #' @param nfam vector (LEN N) containing number of progeny per sire or
 #'   scalar value in case of equal family size
 #' @param pos_chr list (LEN number of chromosomes) of vectors (LEN number of
 #'   markers) of genetic positions in Morgan per chromosome
-#' @param corr logical; \code{TRUE} (default) if output is correlation matrix or
-#'   \code{FALSE} if output is covariance matrix
+#' @param map_fun character string of mapping function used; so far "haldane"
+#'   (default) and "kosambi" are enabled
+#' @param corr logical; \code{FALSE} (default) if output is covariance matrix or
+#'   \code{TRUE} if output is correlation matrix
 #' @return list (LEN 2) of matrix (DIM \eqn{p1} x \eqn{p1}) and vector
 #'   (LEN \eqn{p1}) with \eqn{p1 \le p}
 #' \describe{
@@ -52,12 +55,12 @@
 #'   \url{https://doi.org/10.1101/2019.12.17.879106}
 #' @import foreach
 #' @export
-CovMat <- function(linkDam, haploSire, nfam, pos_chr, corr = T){
-  N <- nrow(haploSire) / 2
+CovMat <- function(linkMat, haploMat, nfam, pos_chr, map_fun = 'haldane', corr = F){
+  N <- nrow(haploMat) / 2
   p <- length(unlist(pos_chr))
 
   # genotypes of sires
-  XSire <- matrix(Haplo2Geno(as.matrix(haploSire)), nrow = N, ncol = p)
+  XSire <- matrix(Haplo2Geno(as.matrix(haploMat)), nrow = N, ncol = p)
 
   # expection of paternally inherited SNP allele
   expectationMat <- matrix(ExpectMat(XSire), nrow = N, ncol = p)
@@ -68,11 +71,11 @@ CovMat <- function(linkDam, haploSire, nfam, pos_chr, corr = T){
 
   # LD of paternally inherited SNP alleles; output is list -> matrices for each family
   indexFam <- NULL
-  linkSire <- foreach(indexFam = 1:N) %do% LDsire(inMat = haploSire, pos_chr, family = fam[[indexFam]], map_fun = "haldane")
+  linkSire <- foreach(indexFam = 1:N) %do% LDsire(inMat = haploMat, pos_chr, family = fam[[indexFam]], map_fun = map_fun)
 
   # Weighted average over paternal half-sib families plus maternal LD
   if(length(nfam) == N) Ns <- nfam else if (length(nfam) == 1) Ns <- rep(nfam, N) else stop("ERROR family size")
-  K <- CovarMatrix(expectationMat, linkDam, linkSire, Ns)
+  K <- CovarMatrix(expectationMat, linkMat, linkSire, Ns)
 
   s <- diag(K)
   id <- s > 1e-6